+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - `info` page showing available screens
+- Target attitude and flight path vector on ADI
 
 ### Changed